Solution 1: Add a basic Action Bar
By using @android:style/Theme.Holo.Light.DarkActionBar, we use the theme
from the video so that you can follow along exactly like the video.

You can also update MainActivity to use this basic Action Bar as well.

Note: This is the original Action Bar used on the video. Android now
uses Toolbar's instead, I will provide another Pull Request which will
allow you to use the newer Toolbar, which looks better and is more
up-to-date. However, things may be slightly different to the video so if
you get stuck, you may have to ask one of us on the forums or find a
solution online.