Solution 2: Add the new Toolbar!
ActionBar's aren't recommended for use anymore so if you want your app
to be up-to-date, merge this commit!

Important: If you merge this commit, there may be some things that are
slightly different to using ActionBar's, like in the video. Therefore,
if you run into any problems on the way, let someone know on the forums
or post a comment on this GitHub and I'll get back to you.

There is quite a bit of code required to do this since our normal
setSupportActionBar() method isn't part of ListActivity but for now, all
you need to know is that it works! Don't try to understand it just yet
:)