Added send button to Action Bar
Added the send buttons to project, added menu_recipients, removed
action_send from menu_main, updated strings.

For some strange reason, when we use the delegated Toolbar, Android
doesn't like app:showAsAction but works fine with android:showAsAction
(This will show as an error in Android Studio, but the code will still
compile fine, however).

Anyhow, use the onCreateOptionsMenu() in RecipientsActivity.java to
inflate the ActionBar or the Toolbar (Works for both) with the send
icon.